(#17456) Initialize settings correctly to use inventory service
With Puppet 3, the settings initialization methods changed when
loading puppet as a library. Without this patch, running the
hiera program with the '-i' flag to search the inventory service
for node information caused a fatal error.

After this change, hiera works correctly with the -i flag
and puppet 3. However it will not work with puppet 2.7.x.